Separate Extraction fns to Extract_Data_To_File.cpp/.h Added PlayerProperties struct to load in player stats Added Init_Player fn for player stats Added Player_Stats to Transform YAML Added DataFiles.h for data related headers Removed LoadValues.cpp/.h Removeed LoadTextures.cpp/.h
