+++ Alpha_Engine/DataFiles.h
@@ -1,0 +1,5 @@
+#pragma once
+#include "pch.h"
+#include "LoadDataFromFile.h"
+#include "ExtractDataToFile.h"
+#include "LevelInitializer.h"
