+++ Alpha_Engine/ExtractDataToFile.h
@@ -1,0 +1,17 @@
+#pragma once
+#include "pch.h"					// For Precompiled Header Files
+#include <string>					// For std::string
+#include <fstream>					// For std::istream/std::ostream
+#include <vector>					// For std::vector
+#include "LoadDataFromFile.h"		// For ObjectTransform & ObjectShape
+
+namespace Extract_Data_To_File{
+
+	using namespace std;
+
+	void Extract_Transform_Data_Out(vector<Load_Data_From_File::ObjectTransform>& vOT, Load_Data_From_File::PlayerProperties sPP, const string level_number);
+	void Extract_Shape_Data_Out(vector<Load_Data_From_File::ObjectShape>& vOS, const string level_number);
+
+
+
+}// END namespace Extract_Data_To_File
