+++ Alpha_Engine/LevelInitializer.h
@@ -1,11 +1,12 @@
 #pragma once
 #include <vector>				// For std::vector
-#include "LoadValues.h"			// For Load_Values::
-#include "LoadDataFromFile.h"	// For Load_Data_From_File::
+#include "DataFiles.h"			// For Load_Data_From_File::
+#include "pch.h"				// For Character
 
 namespace Level_Initializer {
 
 	using namespace std;
 	void Init_Object(vector<Load_Data_From_File::ObjectTransform>& vOT, GameObjects::Object* object[], GameObjects::ObjectInst object_instance[], size_t object_instance_size);
+	void Init_Player(Load_Data_From_File::PlayerProperties* PP, GameObjects::Character& p_player, GameObjects::Object* object, GameObjects::ObjectInst object_instance);
 
 }// END namspace Level_Initializer
